Created TeamProfile component
