+++ src/containers/Pages/AboutPage/TeamProfile/TeamProfile.module.scss
@@ -1,0 +1,95 @@
+.team {
+  width: 65%;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-self: center;
+  justify-content: center;
+}
+
+.magnifyGlassContainer {
+  margin: 1vw;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: flex-start;
+  align-content: flex-start;
+  width: 17vw;
+  height: 500px;
+  z-index: 80;
+  &:hover .imgAvatar {
+    filter: blur(4px);
+    -webkit-filter: blur(4px);
+  }
+  &:hover .magnifyGlassHandle {
+    filter: blur(1px);
+    -webkit-filter: blur(1px);
+  }
+  &:hover .magnifyGlass .logoAvatarContainer .avatarIcon {
+    visibility: visible;
+  }
+  .magnifyGlass {
+    align-self: center;
+    position: relative;
+    border: 0.5vw solid white;
+    border-radius: 50%;
+    width: 15vw;
+    height: 15vw;
+    z-index: 86;
+    .imgAvatar {
+      border-radius: 50%;
+      height: 100%;
+      width: 100%;
+      z-index: 90;
+    }
+
+    .logoAvatarContainer {
+      align-self: center;
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      position: absolute;
+      top: 0%;
+      justify-content: space-evenly;
+      width: 100%;
+      height: 100%;
+      .avatarIcon {
+        height: 3.5vw;
+        width: 3.5vw;
+        z-index: 100;
+        align-self: center;
+        visibility: hidden;
+      }
+    }
+  }
+  .magnifyGlassHandle {
+    height: 0.5vw;
+    width: 25%;
+    background-color: white;
+    transform: rotate(50deg);
+    position: absolute;
+    z-index: 0;
+    left: 75%;
+    border-radius: 15%;
+  }
+
+  &:hover .magnifyGlass {
+    transition: border-width 0.2s ease-in-out;
+    border: 0.7vw solid white;
+  }
+  &:hover .magnifyGlassHandle {
+    transition: border-width 0.2s ease-in-out;
+    border: 0.1vw solid white;
+  }
+  .teamMemberInfo {
+    align-self: flex-end;
+    h4 {
+      text-align: center;
+      margin-bottom: 0.5vw;
+    }
+    p {
+      margin-top: 0;
+    }
+  }
+}
